feat: Improve image generation & project structure
